requirements.txt
datasets>=2.20.0
huggingface_hub>=0.23.0
pyyaml>=6.0.1
tqdm>=4.66.0
openai>=1.30.0